Add files via upload
The objective of this assignment is to develop a responsive, interactive course learning page using HTML and Bootstrap. The page will serve as a platform for users to engage with various learning materials organized into sections and subsections. A central feature of this platform will be a player for course content, accompanied by additional functionalities such as language selection for a global audience.